+++ src/locales/zh-TW.js
@@ -17,6 +17,7 @@ export default {
     saving: '儲存中',
     shareTo: '分享至',
     watchVideo: '觀看原始影片',
+    report: '檢舉',
   },
   post: {
     create: {
@@ -154,6 +155,7 @@ export default {
       followingPosts: '跟著做',
     },
     deletedMessage: '專案未公開或已刪除',
+    report: '檢舉專案',
   },
   home: {
     title: '首頁',
@@ -386,6 +388,10 @@ export default {
     google: '透過 Google 登入',
     facebook: '透過 Facebook 登入',
     loginRequiredMessage: '尚未登入',
+    agreement: {
+      head: '繼續進行即表示你同意',
+      tail: '',
+    },
   },
   searchBar: {
     placeholder: '找專案',
@@ -451,6 +457,11 @@ export default {
     },
     sendFeedbackSuccess: '您的意見已成功送出',
     fileTooLarge: '檔案大小不可超過 %{limit}',
+    reportSuccess: '已檢舉並通知客服人員',
+    reportProject: {
+      title: '檢舉專案',
+      message: '你確定此專案含有不當內容或圖文抄襲嗎？',
+    },
   },
   errors: {
     required: '必須',